refactor: 키보드 제어 유틸을 테마에서 분리
